backend/pyproject.toml
[project]
name = "lnreader-ai-backend"
version = "0.1.0"
description = "Standalone local AI backend for LNReader"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0,<1",
  "httpx>=0.28.0,<1",
  "pydantic-settings>=2.10.0,<3",
  "python-dotenv>=1.1.0,<2",
  "uvicorn[standard]>=0.35.0,<1",
]

[project.scripts]
lnreader-ai-backend = "app.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
